// src/application/dtos/item/active-item-dtos/active-item-response-dto.ts
import { Either } from "../../../../common/helpers/Either";
import DefaultApplicationError from "../../../errors/application/default-application-error";

type ItemData = {
	id: string;
	userId: string;
	name: string;
	description: string;
	price: number;
	category: string;
	status: string;
};

export type ActiveItemResponseDto = Either<DefaultApplicationError, ItemData>;